Update SDK to beta version. Remove/uncomment old tests.
