+++ package.json
@@ -23,7 +23,7 @@
   "author": "karma.run",
   "license": "MIT",
   "dependencies": {
-    "@karma.run/sdk": "^0.5.19",
+    "@karma.run/sdk": "^0.6.0-beta.0",
     "ava": "^1.0.0-beta.4",
     "dotenv": "4.0.0",
     "karma-sdk-typescript": "git+https://github.com/dudagroup/karma-sdk-typescript.git#0.5",
@@ -33,6 +33,6 @@
     "@types/node": "^10.0.0",
     "npm-run-all": "^4.1.2",
     "tslint": "^5.9.1",
-    "typescript": "^3.0.1"
+    "typescript": "^3.2.2"
   }
 }
