+++ src/queries/expressions/dateTime.ts
@@ -1,18 +1,26 @@
-import { buildExpression as build } from '@karma.run/sdk'
 import test from '../_before'
 
+import * as e from '@karma.run/sdk/expression'
+import * as d from '@karma.run/sdk/value'
+
 test('after', async t => {
-  const response = await t.context.exampleQuery('after_0', build(e =>
-    e.after(new Date(0), new Date(1))
-  ))
+  const response = await t.context.exampleQuery('after_0',
+    e.after(
+      e.data(d.dateTime(new Date(0)).toDataConstructor()),
+      e.data(d.dateTime(new Date(1)).toDataConstructor()),
+    )
+  )
 
   t.is(response, false)
 })
 
 test('before', async t => {
-  const response = await t.context.exampleQuery('before_0', build(e =>
-    e.before(new Date(0), new Date(1))
-  ))
+  const response = await t.context.exampleQuery('before_0',
+    e.before(
+      e.data(d.dateTime(new Date(0)).toDataConstructor()),
+      e.data(d.dateTime(new Date(1)).toDataConstructor()),
+    )
+  )
 
   t.is(response, true)
 })
