+++ src/queries/expressions/list.ts
@@ -1,14 +1,13 @@
-import {buildExpression as build} from '@karma.run/sdk'
 import test from '../_before'
 
+import * as e from '@karma.run/sdk/expression'
+import * as d from '@karma.run/sdk/value'
+
 test('concatLists', async t => {
-  const response = await t.context.exampleQuery(
-    'concatLists_0',
-    build(e =>
-      e.concatLists(
-        e.data(d => d.list(d.int8(1), d.int8(2), d.int8(3))),
-        e.data(d => d.list(d.int8(4), d.int8(5), d.int8(6)))
-      )
+  const response = await t.context.exampleQuery('concatLists_0',
+    e.concatLists(
+      e.data(d.list([d.int8(1), d.int8(2), d.int8(3)]).toDataConstructor()),
+      e.data(d.list([d.int8(4), d.int8(5), d.int8(6)]).toDataConstructor())
     )
   )
 
@@ -18,10 +17,14 @@ test('concatLists', async t => {
 test('filterList', async t => {
   const response = await t.context.exampleQuery(
     'filterList_0',
-    build(e =>
-      e.filterList(e.data(d => d.list(d.int8(3), d.int8(2), d.int8(1))), (_, value) =>
-        e.gtInt8(value, e.data(d => d.int8(2)))
-      )
+    e.filterList(
+      e.data(d.list([d.int8(3), d.int8(2), d.int8(1)]).toDataConstructor()),
+      (index, value) => {
+        return e.gtInt8(
+          value,
+          e.data(d.int8(2).toDataConstructor())
+        )
+      }
     )
   )
 
@@ -29,38 +32,47 @@ test('filterList', async t => {
 })
 
 test('first', async t => {
-  const response = await t.context.exampleQuery(
-    'first_0',
-    build(e => e.first(e.data(d => d.list(d.int8(3), d.int8(2), d.int8(1)))))
+  const response = await t.context.exampleQuery('first_0',
+    e.first(e.data(
+      d.list([d.int8(3), d.int8(2), d.int8(1)]).toDataConstructor()
+    ))
   )
 
   t.is(response, 3)
 })
 
 test('inList', async t => {
-  let response = await t.context.exampleQuery(
-    'inList_0',
-    build(e =>
-      e.inList(e.data(d => d.list(d.string('foo'), d.int8(10), d.int8(15))), e.string('foo'))
-    )
+  let response = await t.context.exampleQuery('inList_0',
+
+    e.inList(e.data(
+      d.list([
+        d.string('foo'),
+        d.int8(10),
+        d.int8(15)
+      ]).toDataConstructor()
+    ), e.string('foo'))
   )
 
   t.true(response)
 
-  response = await t.context.exampleQuery(
-    'inList_1',
-    build(e =>
-      e.inList(e.data(d => d.list(d.string('foo'), d.int8(10), d.int8(15))), e.string('bar'))
-    )
+  response = await t.context.exampleQuery('inList_1',
+    e.inList(e.data(
+      d.list([
+        d.string('foo'),
+        d.int8(10),
+        d.int8(15)
+      ]).toDataConstructor()
+    ), e.string('bar'))
   )
 
   t.false(response)
 })
 
 test('mapList', async t => {
-  const response = await t.context.exampleQuery(
-    'mapList_0',
-    build(e => e.mapList(e.all(e.tag('_tag')), (_, value) => e.field('tag', value)))
+  const response = await t.context.exampleQuery('mapList_0',
+    e.mapList(e.all(e.tag('_tag')),
+      (_, value) => e.field('tag', value)
+    )
   )
 
   t.deepEqual(
@@ -70,31 +82,29 @@ test('mapList', async t => {
 })
 
 test('length', async t => {
-  const response = await t.context.exampleQuery(
-    'length_0',
-    build(e => e.length(e.all(e.tag('_tag'))))
+  const response = await t.context.exampleQuery('length_0',
+    e.length(e.all(e.tag('_tag')))
   )
 
   t.deepEqual(response, 6)
 })
 
 test('memSort', async t => {
-  const response = await t.context.exampleQuery(
-    'memSort_0',
-    build(e => e.memSort(e.data(d => d.list(d.int8(2), d.int8(8), d.int8(4))), value => value))
+  const response = await t.context.exampleQuery('memSort_0',
+    e.memSort(e.data(d.list([d.int8(2), d.int8(8), d.int8(4)]).toDataConstructor()), value => value)
   )
 
   t.deepEqual(response, [2, 4, 8])
 })
 
-
 test('memSortFunction', async t => {
-  const response = await t.context.exampleQuery(
-    'memSortFunction_0',
-    build(e =>
-      e.memSortFunction(e.data(d => d.list(d.int8(2), d.int8(8), d.int8(4))), (valueA, valueB) =>
+  const response = await t.context.exampleQuery('memSortFunction_0',
+    e.memSortFunction(e.data(
+      d.list([
+        d.int8(2), d.int8(8), d.int8(4)
+      ]).toDataConstructor()
+      ), (valueA, valueB) =>
         e.gtInt8(valueA, valueB)
-      )
     )
   )
 
@@ -102,31 +112,31 @@ test('memSortFunction', async t => {
 })
 
 test('reverseList', async t => {
-  const response = await t.context.exampleQuery(
-    'reverseList_0',
-    build(e => e.reverseList(e.data(d => d.list(d.int8(1), d.int8(2), d.int8(3)))))
+  const response = await t.context.exampleQuery('reverseList_0',
+    e.reverseList(e.data(d.list([d.int8(1), d.int8(2), d.int8(3)]).toDataConstructor()))
   )
 
   t.deepEqual(response, [3, 2, 1])
 })
+
 test('slice', async t => {
-  const response = await t.context.exampleQuery(
-    'slice_0',
-    build(e => e.slice(e.data(d => d.list(d.int8(1), d.int8(2), d.int8(3))), 1, 2))
+  const response = await t.context.exampleQuery('slice_0',
+    e.slice(e.data(d.list([d.int8(1), d.int8(2), d.int8(3)]).toDataConstructor()), 1, 2)
   )
 
   t.deepEqual(response, [2, 3])
 })
 
 test('reduceList', async t => {
-  const response = await t.context.exampleQuery(
-    'reduceList_0',
-    build(e =>
-      e.reduceList(
-        e.data(d => d.list(d.int8(5), d.int8(10), d.int8(15))),
-        e.int8(0),
-        (value, nextValue) => e.addInt8(value, nextValue)
-      )
+  const response = await t.context.exampleQuery('reduceList_0',
+    e.reduceList(
+      e.data(
+        d.list([
+          d.int8(5), d.int8(10), d.int8(15)
+        ]).toDataConstructor()
+      ),
+      e.int8(0),
+      (value, nextValue) => e.addInt8(value, nextValue)
     )
   )
 
@@ -134,22 +144,24 @@ test('reduceList', async t => {
 })
 
 test('leftFoldList', async t => {
-  const response = await t.context.exampleQuery(
-    'leftFoldList_0',
-    build(e =>
-      e.leftFoldList(
-        e.data(d => d.list(d.string('bar'), d.string('baz'))),
-        e.data(d => d.struct({value: d.string('foo')})),
-        (aggregator, value) =>
-          e.setField(
-            'value',
-            e.joinStrings(
-              e.string(' '),
-              e.data(d => d.list(d.expr(e.field('value', aggregator)), d.expr(value)))
-            ),
-            aggregator
-          )
-      )
+  const dc = e.DataContext
+
+  const response = await t.context.exampleQuery('leftFoldList_0',
+    e.leftFoldList(
+      e.data(d.list([d.string('bar'), d.string('baz')]).toDataConstructor()),
+      e.data(d.struct({value: d.string('foo')}).toDataConstructor()),
+      (aggregator, value) =>
+        e.setField(
+          'value',
+          e.joinStrings(
+            e.data(dc.list([
+              dc.expr(e.field('value', aggregator)),
+              dc.expr(value)
+            ])),
+            e.string(' '),
+          ),
+          aggregator
+        )
     )
   )
 
@@ -157,22 +169,24 @@ test('leftFoldList', async t => {
 })
 
 test('rightFoldList', async t => {
-  const response = await t.context.exampleQuery(
-    'rightFoldList_0',
-    build(e =>
-      e.rightFoldList(
-        e.data(d => d.list(d.string('bar'), d.string('baz'))),
-        e.data(d => d.struct({value: d.string('foo')})),
-        (aggregator, value) =>
-          e.setField(
-            'value',
-            e.joinStrings(
-              e.string(' '),
-              e.data(d => d.list(d.expr(e.field('value', aggregator)), d.expr(value)))
-            ),
-            aggregator
-          )
-      )
+  const dc = e.DataContext
+
+  const response = await t.context.exampleQuery('rightFoldList_0',
+    e.rightFoldList(
+      e.data(d.list([d.string('bar'), d.string('baz')]).toDataConstructor()),
+      e.data(d.struct({value: d.string('foo')}).toDataConstructor()),
+      (aggregator, value) =>
+        e.setField(
+          'value',
+          e.joinStrings(
+            e.data(dc.list([
+              dc.expr(e.field('value', aggregator)),
+              dc.expr(value),
+            ])),
+            e.string(' '),
+          ),
+          aggregator
+        )
     )
   )
