+++ src/queries/expressions/numeric.ts
@@ -1,12 +1,17 @@
-import {buildExpression as build} from '@karma.run/sdk'
 import test from '../_before'
+
+import * as e from '@karma.run/sdk/expression'
+import * as d from '@karma.run/sdk/value'
 
 // Addition
 // ========
 
 test('addFloat', async t => {
   const response = await t.context.exampleQuery('addFloat_0',
-    build(e => e.addFloat(0.5, 0.5))
+    e.addFloat(
+      e.data(d.float(0.5).toDataConstructor()),
+      e.data(d.float(0.5).toDataConstructor())
+    )
   )
 
   t.is(response, 1)
@@ -14,7 +19,10 @@ test('addFloat', async t => {
 
 test('addInt8', async t => {
   const response = await t.context.exampleQuery('addInt8_0',
-    build(e => e.addInt8(-128, 127))
+    e.addInt8(
+      e.data(d.int8(-128).toDataConstructor()),
+      e.data(d.int8(127).toDataConstructor())
+    )
   )
 
   t.is(response, -1)
@@ -22,7 +30,10 @@ test('addInt8', async t => {
 
 test('addInt16', async t => {
   const response = await t.context.exampleQuery('addInt16_0',
-    build(e => e.addInt16(-32_768, 32_767))
+    e.addInt16(
+      e.data(d.int16(-32_768).toDataConstructor()),
+      e.data(d.int16(32_767).toDataConstructor())
+    )
   )
 
   t.is(response, -1)
@@ -30,7 +41,10 @@ test('addInt16', async t => {
 
 test('addInt32', async t => {
   const response = await t.context.exampleQuery('addInt32_0',
-    build(e => e.addInt32(-2_147_483_648, 2_147_483_647))
+    e.addInt32(
+      e.data(d.int32(-2_147_483_648).toDataConstructor()),
+      e.data(d.int32(2_147_483_647).toDataConstructor())
+    )
   )
 
   t.is(response, -1)
@@ -38,7 +52,10 @@ test('addInt32', async t => {
 
 test('addInt64', async t => {
   const response = await t.context.exampleQuery('addInt64_0',
-    build(e => e.addInt64(Number.MIN_SAFE_INTEGER, Number.MAX_SAFE_INTEGER))
+    e.addInt64(
+      e.data(d.int64(Number.MIN_SAFE_INTEGER).toDataConstructor()),
+      e.data(d.int64(Number.MAX_SAFE_INTEGER).toDataConstructor())
+    )
   )
 
   t.is(response, 0)
@@ -46,7 +63,10 @@ test('addInt64', async t => {
 
 test('addUint8', async t => {
   const response = await t.context.exampleQuery('addUint8_0',
-    build(e => e.addUint8(e.uint8(128), e.uint8(127)))
+    e.addUint8(
+      e.data(d.uint8(128).toDataConstructor()),
+      e.data(d.uint8(127).toDataConstructor())
+    )
   )
 
   t.is(response, 255)
@@ -54,7 +74,10 @@ test('addUint8', async t => {
 
 test('addUint16', async t => {
   const response = await t.context.exampleQuery('addUint16_0',
-    build(e => e.addUint16(e.uint16(32_768), e.uint16(32_767)))
+    e.addUint16(
+      e.data(d.uint16(32_768).toDataConstructor()),
+      e.data(d.uint16(32_767).toDataConstructor())
+    )
   )
 
   t.is(response, 65_535)
@@ -62,7 +85,10 @@ test('addUint16', async t => {
 
 test('addUint32', async t => {
   const response = await t.context.exampleQuery('addUint32_0',
-    build(e => e.addUint32(e.uint32(2_147_483_648), e.uint32(2_147_483_647)))
+    e.addUint32(
+      e.data(d.uint32(2_147_483_648).toDataConstructor()),
+      e.data(d.uint32(2_147_483_647).toDataConstructor())
+    )
   )
 
   t.is(response, 4_294_967_295)
@@ -70,7 +96,10 @@ test('addUint32', async t => {
 
 test('addUint64', async t => {
   const response = await t.context.exampleQuery('addUint64_0',
-    build(e => e.addUint64(e.uint64(1), e.uint64(Number.MAX_SAFE_INTEGER - 1)))
+    e.addUint64(
+      e.data(d.uint64(1).toDataConstructor()),
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER - 1).toDataConstructor())
+    )
   )
 
   t.is(response, Number.MAX_SAFE_INTEGER)
@@ -81,7 +110,10 @@ test('addUint64', async t => {
 
 test('subFloat', async t => {
   const response = await t.context.exampleQuery('subFloat_0',
-    build(e => e.subFloat(e.float(0.5), e.float(0.5)))
+    e.subFloat(
+      e.data(d.float(0.5).toDataConstructor()),
+      e.data(d.float(0.5).toDataConstructor())
+    )
   )
 
   t.is(response, 0)
@@ -89,7 +121,10 @@ test('subFloat', async t => {
 
 test('subInt8', async t => {
   const response = await t.context.exampleQuery('subInt8_0',
-    build(e => e.subInt8(-128, 127))
+    e.subInt8(
+      e.data(d.int8(-128).toDataConstructor()),
+      e.data(d.int8(127).toDataConstructor())
+    )
   )
 
   t.is(response, 1)
@@ -97,7 +132,10 @@ test('subInt8', async t => {
 
 test('subInt16', async t => {
   const response = await t.context.exampleQuery('subInt16_0',
-    build(e => e.subInt16(-32_768, 32_767))
+    e.subInt16(
+      e.data(d.int16(-32_768).toDataConstructor()),
+      e.data(d.int16(32_767).toDataConstructor())
+    )
   )
 
   t.is(response, 1)
@@ -105,7 +143,10 @@ test('subInt16', async t => {
 
 test('subInt32', async t => {
   const response = await t.context.exampleQuery('subInt32_0',
-    build(e => e.subInt32(-2_147_483_648, 2_147_483_647))
+    e.subInt32(
+      e.data(d.int32(-2_147_483_648).toDataConstructor()),
+      e.data(d.int32(2_147_483_647).toDataConstructor())
+    )
   )
 
   t.is(response, 1)
@@ -113,7 +154,10 @@ test('subInt32', async t => {
 
 test('subInt64', async t => {
   const response = await t.context.exampleQuery('subInt64_0',
-    build(e => e.subInt64(0, Number.MAX_SAFE_INTEGER))
+    e.subInt64(
+      e.data(d.int64(0).toDataConstructor()),
+      e.data(d.int64(Number.MAX_SAFE_INTEGER).toDataConstructor())
+    )
   )
 
   t.is(response, -Number.MAX_SAFE_INTEGER)
@@ -121,7 +165,10 @@ test('subInt64', async t => {
 
 test('subUint8', async t => {
   const response = await t.context.exampleQuery('subUint8_0',
-    build(e => e.subUint8(e.uint8(128), e.uint8(127)))
+    e.subUint8(
+      e.data(d.uint8(128).toDataConstructor()),
+      e.data(d.uint8(127).toDataConstructor())
+    )
   )
 
   t.is(response, 1)
@@ -129,7 +176,10 @@ test('subUint8', async t => {
 
 test('subUint16', async t => {
   const response = await t.context.exampleQuery('subUint16_0',
-    build(e => e.subUint16(e.uint16(32_768), e.uint16(32_767)))
+    e.subUint16(
+      e.data(d.uint16(32_768).toDataConstructor()),
+      e.data(d.uint16(32_767).toDataConstructor()),
+    )
   )
 
   t.is(response, 1)
@@ -137,7 +187,10 @@ test('subUint16', async t => {
 
 test('subUint32', async t => {
   const response = await t.context.exampleQuery('subUint32_0',
-    build(e => e.subUint32(e.uint32(2_147_483_648), e.uint32(2_147_483_647)))
+    e.subUint32(
+      e.data(d.uint32(2_147_483_648).toDataConstructor()),
+      e.data(d.uint32(2_147_483_647).toDataConstructor()),
+    )
   )
 
   t.is(response, 1)
@@ -145,7 +198,10 @@ test('subUint32', async t => {
 
 test('subUint64', async t => {
   const response = await t.context.exampleQuery('subUint64_0',
-    build(e => e.subUint64(e.uint64(Number.MAX_SAFE_INTEGER), e.uint64(Number.MAX_SAFE_INTEGER - 1)))
+    e.subUint64(
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER - 1).toDataConstructor())
+    )
   )
 
   t.is(response, 1)
@@ -156,7 +212,10 @@ test('subUint64', async t => {
 
 test('divFloat', async t => {
   const response = await t.context.exampleQuery('divFloat_0',
-    build(e => e.divFloat(e.float(1), e.float(2)))
+    e.divFloat(
+      e.data(d.float(1).toDataConstructor()),
+      e.data(d.float(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 0.5)
@@ -164,7 +223,10 @@ test('divFloat', async t => {
 
 test('divInt8', async t => {
   const response = await t.context.exampleQuery('divInt8_0',
-    build(e => e.divInt8(127, 2))
+    e.divInt8(
+      e.data(d.int8(127).toDataConstructor()),
+      e.data(d.int8(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 63)
@@ -172,7 +234,10 @@ test('divInt8', async t => {
 
 test('divInt16', async t => {
   const response = await t.context.exampleQuery('divInt16_0',
-    build(e => e.divInt16(32_767, 2))
+    e.divInt16(
+      e.data(d.int16(32_767).toDataConstructor()),
+      e.data(d.int16(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 16_383)
@@ -180,7 +245,10 @@ test('divInt16', async t => {
 
 test('divInt32', async t => {
   const response = await t.context.exampleQuery('divInt32_0',
-    build(e => e.divInt32(2_147_483_647, 2))
+    e.divInt32(
+      e.data(d.int32(2_147_483_647).toDataConstructor()),
+      e.data(d.int32(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 1_073_741_823)
@@ -188,7 +256,10 @@ test('divInt32', async t => {
 
 test('divInt64', async t => {
   const response = await t.context.exampleQuery('divInt64_0',
-    build(e => e.divInt64(Number.MAX_SAFE_INTEGER, 2))
+    e.divInt64(
+      e.data(d.int64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+      e.data(d.int64(2).toDataConstructor()),
+    )
   )
 
   t.is(response, Math.floor(Number.MAX_SAFE_INTEGER / 2))
@@ -196,7 +267,10 @@ test('divInt64', async t => {
 
 test('divUint8', async t => {
   const response = await t.context.exampleQuery('divUint8_0',
-    build(e => e.divUint8(e.uint8(255), e.uint8(2)))
+    e.divUint8(
+      e.data(d.uint8(255).toDataConstructor()),
+      e.data(d.uint8(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 127)
@@ -204,7 +278,10 @@ test('divUint8', async t => {
 
 test('divUint16', async t => {
   const response = await t.context.exampleQuery('divUint16_0',
-    build(e => e.divUint16(e.uint16(65_535), e.uint16(2)))
+    e.divUint16(
+      e.data(d.uint16(65_535).toDataConstructor()),
+      e.data(d.uint16(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 32_767)
@@ -212,7 +289,10 @@ test('divUint16', async t => {
 
 test('divUint32', async t => {
   const response = await t.context.exampleQuery('divUint32_0',
-    build(e => e.divUint32(e.uint32(4_294_967_295), e.uint32(2)))
+    e.divUint32(
+      e.data(d.uint32(4_294_967_295).toDataConstructor()),
+      e.data(d.uint32(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 2_147_483_647)
@@ -220,7 +300,10 @@ test('divUint32', async t => {
 
 test('divUint64', async t => {
   const response = await t.context.exampleQuery('divUint64_0',
-    build(e => e.divUint64(e.uint64(Number.MAX_SAFE_INTEGER), e.uint64(2)))
+    e.divUint64(
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+      e.data(d.uint64(2).toDataConstructor()),
+    )
   )
 
   t.is(response, Math.floor(Number.MAX_SAFE_INTEGER / 2))
@@ -231,7 +314,10 @@ test('divUint64', async t => {
 
 test('mulFloat', async t => {
   const response = await t.context.exampleQuery('mulFloat_0',
-    build(e => e.mulFloat(e.float(0.5), e.float(2)))
+    e.mulFloat(
+      e.data(d.float(0.5).toDataConstructor()),
+      e.data(d.float(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 1)
@@ -239,7 +325,10 @@ test('mulFloat', async t => {
 
 test('mulInt8', async t => {
   const response = await t.context.exampleQuery('mulInt8_0',
-    build(e => e.mulInt8(63, 2))
+    e.mulInt8(
+      e.data(d.int8(63).toDataConstructor()),
+      e.data(d.int8(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 127 - 1)
@@ -247,7 +336,10 @@ test('mulInt8', async t => {
 
 test('mulInt16', async t => {
   const response = await t.context.exampleQuery('mulInt16_0',
-    build(e => e.mulInt16(16_383, 2))
+    e.mulInt16(
+      e.data(d.int16(16_383).toDataConstructor()),
+      e.data(d.int16(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 32_767 - 1)
@@ -255,7 +347,10 @@ test('mulInt16', async t => {
 
 test('mulInt32', async t => {
   const response = await t.context.exampleQuery('mulInt32_0',
-    build(e => e.mulInt32(1_073_741_823, 2))
+    e.mulInt32(
+      e.data(d.int32(1_073_741_823).toDataConstructor()),
+      e.data(d.int32(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 2_147_483_647 - 1)
@@ -263,7 +358,10 @@ test('mulInt32', async t => {
 
 test('mulInt64', async t => {
   const response = await t.context.exampleQuery('mulInt64_0',
-    build(e => e.mulInt64(Math.floor(Number.MAX_SAFE_INTEGER / 2), 2))
+    e.mulInt64(
+      e.data(d.int64(Math.floor(Number.MAX_SAFE_INTEGER / 2)).toDataConstructor()),
+      e.data(d.int64(2).toDataConstructor()),
+    )
   )
 
   t.is(response, Number.MAX_SAFE_INTEGER - 1)
@@ -271,7 +369,10 @@ test('mulInt64', async t => {
 
 test('mulUint8', async t => {
   const response = await t.context.exampleQuery('mulUint8_0',
-    build(e => e.mulUint8(e.uint8(127), e.uint8(2)))
+    e.mulUint8(
+      e.data(d.uint8(127).toDataConstructor()),
+      e.data(d.uint8(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 255 - 1)
@@ -279,7 +380,10 @@ test('mulUint8', async t => {
 
 test('mulUint16', async t => {
   const response = await t.context.exampleQuery('mulUint16_0',
-    build(e => e.mulUint16(e.uint16(32_767), e.uint16(2)))
+    e.mulUint16(
+      e.data(d.uint16(32_767).toDataConstructor()),
+      e.data(d.uint16(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 65_535 - 1)
@@ -287,7 +391,10 @@ test('mulUint16', async t => {
 
 test('mulUint32', async t => {
   const response = await t.context.exampleQuery('mulUint32_0',
-    build(e => e.mulUint32(e.uint32(2_147_483_647), e.uint32(2)))
+    e.mulUint32(
+      e.data(d.uint32(2_147_483_647).toDataConstructor()),
+      e.data(d.uint32(2).toDataConstructor()),
+    )
   )
 
   t.is(response, 4_294_967_295 - 1)
@@ -295,9 +402,11 @@ test('mulUint32', async t => {
 
 test('mulUint64', async t => {
   const response = await t.context.exampleQuery('mulUint64_0',
-    build(e => e.mulUint64(e.uint64(Math.floor(Number.MAX_SAFE_INTEGER / 2)), e.uint64(2)))
+    e.mulUint64(
+      e.data(d.uint64(Math.floor(Number.MAX_SAFE_INTEGER / 2)).toDataConstructor()),
+      e.data(d.uint64(2).toDataConstructor()),
+    )
   )
-
   t.is(response, Number.MAX_SAFE_INTEGER - 1)
 })
 
@@ -306,7 +415,10 @@ test('mulUint64', async t => {
 
 test('ltFloat', async t => {
   const response = await t.context.exampleQuery('ltFloat_0',
-    build(e => e.ltFloat(e.float(Number.MIN_VALUE), e.float(Number.MAX_VALUE)))
+    e.ltFloat(
+      e.data(d.float(Number.MIN_VALUE).toDataConstructor()),
+      e.data(d.float(Number.MAX_VALUE).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -314,7 +426,10 @@ test('ltFloat', async t => {
 
 test('ltInt8', async t => {
   const response = await t.context.exampleQuery('ltInt8_0',
-    build(e => e.ltInt8(-128, 127))
+    e.ltInt8(
+      e.data(d.int8(-128).toDataConstructor()),
+      e.data(d.int8(127).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -322,7 +437,10 @@ test('ltInt8', async t => {
 
 test('ltInt16', async t => {
   const response = await t.context.exampleQuery('ltInt16_0',
-    build(e => e.ltInt16(-32_768, 32_767))
+    e.ltInt16(
+      e.data(d.int16(-32_768).toDataConstructor()),
+      e.data(d.int16(32_767).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -330,7 +448,10 @@ test('ltInt16', async t => {
 
 test('ltInt32', async t => {
   const response = await t.context.exampleQuery('ltInt32_0',
-    build(e => e.ltInt32(-2_147_483_648, 2_147_483_647))
+    e.ltInt32(
+      e.data(d.int32(-2_147_483_648).toDataConstructor()),
+      e.data(d.int32(2_147_483_647).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -338,7 +459,10 @@ test('ltInt32', async t => {
 
 test('ltInt64', async t => {
   const response = await t.context.exampleQuery('ltInt64_0',
-    build(e => e.ltInt64(Number.MIN_SAFE_INTEGER, Number.MAX_SAFE_INTEGER))
+    e.ltInt64(
+      e.data(d.int64(Number.MIN_SAFE_INTEGER).toDataConstructor()),
+      e.data(d.int64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -346,7 +470,10 @@ test('ltInt64', async t => {
 
 test('ltUint8', async t => {
   const response = await t.context.exampleQuery('ltUint8_0',
-    build(e => e.ltUint8(e.uint8(0), e.uint8(255)))
+    e.ltUint8(
+      e.data(d.uint8(0).toDataConstructor()),
+      e.data(d.uint8(255).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -354,7 +481,10 @@ test('ltUint8', async t => {
 
 test('ltUint16', async t => {
   const response = await t.context.exampleQuery('ltUint16_0',
-    build(e => e.ltUint16(e.uint16(0), e.uint16(65_535)))
+    e.ltUint16(
+      e.data(d.uint16(0).toDataConstructor()),
+      e.data(d.uint16(65_535).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -362,7 +492,10 @@ test('ltUint16', async t => {
 
 test('ltUint32', async t => {
   const response = await t.context.exampleQuery('ltUint32_0',
-    build(e => e.ltUint32(e.uint32(0), e.uint32(4_294_967_295)))
+    e.ltUint32(
+      e.data(d.uint32(0).toDataConstructor()),
+      e.data(d.uint32(4_294_967_295).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -370,7 +503,10 @@ test('ltUint32', async t => {
 
 test('ltUint64', async t => {
   const response = await t.context.exampleQuery('ltUint64_0',
-    build(e => e.ltUint64(e.uint64(0), e.uint64(Number.MAX_SAFE_INTEGER)))
+    e.ltUint64(
+      e.data(d.uint64(0).toDataConstructor()),
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+    )
   )
 
   t.is(response, true)
@@ -381,7 +517,10 @@ test('ltUint64', async t => {
 
 test('gtFloat', async t => {
   const response = await t.context.exampleQuery('gtFloat_0',
-    build(e => e.gtFloat(e.float(Number.MIN_VALUE), e.float(Number.MAX_VALUE)))
+    e.gtFloat(
+      e.data(d.float(Number.MIN_VALUE).toDataConstructor()),
+      e.data(d.float(Number.MAX_VALUE).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -389,7 +528,10 @@ test('gtFloat', async t => {
 
 test('gtInt8', async t => {
   const response = await t.context.exampleQuery('gtInt8_0',
-    build(e => e.gtInt8(-128, 127))
+    e.gtInt8(
+      e.data(d.int8(-128).toDataConstructor()),
+      e.data(d.int8(127).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -397,7 +539,10 @@ test('gtInt8', async t => {
 
 test('gtInt16', async t => {
   const response = await t.context.exampleQuery('gtInt16_0',
-    build(e => e.gtInt16(-32_768, 32_767))
+    e.gtInt16(
+      e.data(d.int16(-32_768).toDataConstructor()),
+      e.data(d.int16(32_767).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -405,7 +550,10 @@ test('gtInt16', async t => {
 
 test('gtInt32', async t => {
   const response = await t.context.exampleQuery('gtInt32_0',
-    build(e => e.gtInt32(-2_147_483_648, 2_147_483_647))
+    e.gtInt32(
+      e.data(d.int32(-2_147_483_648).toDataConstructor()),
+      e.data(d.int32(2_147_483_647).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -413,7 +561,10 @@ test('gtInt32', async t => {
 
 test('gtInt64', async t => {
   const response = await t.context.exampleQuery('gtInt64_0',
-    build(e => e.gtInt64(Number.MIN_SAFE_INTEGER, Number.MAX_SAFE_INTEGER))
+    e.gtInt64(
+      e.data(d.int64(Number.MIN_SAFE_INTEGER).toDataConstructor()),
+      e.data(d.int64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -421,7 +572,10 @@ test('gtInt64', async t => {
 
 test('gtUint8', async t => {
   const response = await t.context.exampleQuery('gtUint8_0',
-    build(e => e.gtUint8(e.uint8(0), e.uint8(255)))
+    e.gtUint8(
+      e.data(d.uint8(0).toDataConstructor()),
+      e.data(d.uint8(255).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -429,7 +583,10 @@ test('gtUint8', async t => {
 
 test('gtUint16', async t => {
   const response = await t.context.exampleQuery('gtUint16_0',
-    build(e => e.gtUint16(e.uint16(0), e.uint16(65_535)))
+    e.gtUint16(
+      e.data(d.uint16(0).toDataConstructor()),
+      e.data(d.uint16(65_535).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -437,7 +594,10 @@ test('gtUint16', async t => {
 
 test('gtUint32', async t => {
   const response = await t.context.exampleQuery('gtUint32_0',
-    build(e => e.gtUint32(e.uint32(0), e.uint32(4_294_967_295)))
+    e.gtUint32(
+      e.data(d.uint32(0).toDataConstructor()),
+      e.data(d.uint32(4_294_967_295).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -445,7 +605,10 @@ test('gtUint32', async t => {
 
 test('gtUint64', async t => {
   const response = await t.context.exampleQuery('gtUint64_0',
-    build(e => e.gtUint64(e.uint64(0), e.uint64(Number.MAX_SAFE_INTEGER)))
+    e.gtUint64(
+      e.data(d.uint64(0).toDataConstructor()),
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+    )
   )
 
   t.is(response, false)
@@ -456,7 +619,9 @@ test('gtUint64', async t => {
 
 test('toFloat', async t => {
   const response = await t.context.exampleQuery('toFloat_0',
-    build(e => e.toFloat(e.uint64(Number.MAX_SAFE_INTEGER)))
+    e.toFloat(
+      e.data(d.uint64(Number.MAX_SAFE_INTEGER).toDataConstructor()),
+    )
   )
 
   t.is(response, Number.MAX_SAFE_INTEGER)
@@ -464,7 +629,9 @@ test('toFloat', async t => {
 
 test('toInt8', async t => {
   const response = await t.context.exampleQuery('toInt8_0',
-    build(e => e.toInt8(e.float(127.5)))
+    e.toInt8(
+      e.data(d.float(127.5).toDataConstructor()),
+    )
   )
 
   t.is(response, 127)
@@ -472,7 +639,9 @@ test('toInt8', async t => {
 
 test('toInt16', async t => {
   const response = await t.context.exampleQuery('toInt16_0',
-    build(e => e.toInt16(e.float(32_767.5)))
+    e.toInt16(
+      e.data(d.float(32_767.5).toDataConstructor()),
+    )
   )
 
   t.is(response, 32_767)
@@ -480,7 +649,9 @@ test('toInt16', async t => {
 
 test('toInt32', async t => {
   const response = await t.context.exampleQuery('toInt32_0',
-    build(e => e.toInt32(e.float(2_147_483_647.5)))
+    e.toInt32(
+      e.data(d.float(2_147_483_647.5).toDataConstructor()),
+    )
   )
 
   t.is(response, 2_147_483_647)
@@ -488,7 +659,9 @@ test('toInt32', async t => {
 
 test('toInt64', async t => {
   const response = await t.context.exampleQuery('toInt64_0',
-    build(e => e.toInt64(e.float(Number.MAX_SAFE_INTEGER - 0.5)))
+    e.toInt64(
+      e.data(d.float(Number.MAX_SAFE_INTEGER - 0.5).toDataConstructor()),
+    )
   )
 
   t.is(response, Number.MAX_SAFE_INTEGER - 1)
@@ -496,7 +669,9 @@ test('toInt64', async t => {
 
 test('toUint8', async t => {
   const response = await t.context.exampleQuery('toUint8_0',
-    build(e => e.toUint8(e.float(127.5)))
+    e.toUint8(
+      e.data(d.float(127.5).toDataConstructor()),
+    )
   )
 
   t.is(response, 127)
@@ -504,7 +679,9 @@ test('toUint8', async t => {
 
 test('toUint16', async t => {
   const response = await t.context.exampleQuery('toUint16_0',
-    build(e => e.toUint16(e.float(32_767.5)))
+    e.toUint16(
+      e.data(d.float(32_767.5).toDataConstructor()),
+    )
   )
 
   t.is(response, 32_767)
@@ -512,7 +689,9 @@ test('toUint16', async t => {
 
 test('toUint32', async t => {
   const response = await t.context.exampleQuery('toUint32_0',
-    build(e => e.toUint32(e.float(2_147_483_647.5)))
+    e.toUint32(
+      e.data(d.float(2_147_483_647.5).toDataConstructor()),
+    )
   )
 
   t.is(response, 2_147_483_647)
@@ -520,7 +699,9 @@ test('toUint32', async t => {
 
 test('toUint64', async t => {
   const response = await t.context.exampleQuery('toUint64_0',
-    build(e => e.toUint64(e.float(Number.MAX_SAFE_INTEGER - 0.5)))
+    e.toUint64(
+      e.data(d.float(Number.MAX_SAFE_INTEGER - 0.5).toDataConstructor()),
+    )
   )
 
   t.is(response, Number.MAX_SAFE_INTEGER - 1)
