+++ src/queries/expressions/semantic.ts
@@ -1,23 +1,23 @@
-import {model as m, expression as e} from '@karma.run/sdk'
 import test from '../_before'
 
+import * as e from '@karma.run/sdk/expression'
+import * as m from '@karma.run/sdk/model'
+
+import {isRef} from '../utility'
+
 test('annotation', async t => {
-  const response = await t.context.exampleQuery('annotation_0',
-    e.data(
-      m.struct({
-        annotatedKey: m.annotation("custom annotation string", m.string())
-      })
+  const metaRef = await t.context.adminSession.getMetaModelRef()
+
+  const model = m.struct({
+    annotatedKey: m.annotation('custom annotation string', m.string),
+    myString: m.string,
+  })
+
+  const response = await t.context.exampleQuery('create_0',
+    e.create(e.tag('_model'),
+      arg => e.data(model.toValue(metaRef.id).toDataConstructor())
     )
   )
 
-  t.deepEqual(response, {
-    "struct": {
-      "annotatedKey": {
-        "annotation": {
-          "model": {"string": {}},
-          "value": "custom annotation string"
-        }
-      }
-    }
-  })
+  t.true(isRef(response))
 })
