+++ src/queries/expressions/set.ts
@@ -1,17 +1,21 @@
-import { buildExpression as build, isRef } from '@karma.run/sdk'
 import test from '../_before'
 
+import * as e from '@karma.run/sdk/expression'
+import * as d from '@karma.run/sdk/value'
+
+import {isRef} from '../utility'
+
 test('mapSet', async t => {
-  const response = await t.context.exampleQuery('mapSet_0', build(e =>
+  const response = await t.context.exampleQuery('mapSet_0',
     e.mapSet(
-      e.data(d => d.set(
+      e.data(d.set([
         d.string('_tag'),
         d.string('_tag'),
-        d.string('_user')
-      )),
+        d.string('_user'),
+      ]).toDataConstructor()),
       (value) => e.tag(value)
     )
-  ))
+  )
 
   t.true(Array.isArray(response))
   t.true(isRef(response[0]))
