+++ src/queries/expressions/struct.ts
@@ -1,48 +1,56 @@
-import { buildExpression as build } from '@karma.run/sdk'
 import test from '../_before'
+
+import * as e from '@karma.run/sdk/expression'
+import * as d from '@karma.run/sdk/value'
+
+import {isRef} from '../utility'
 
 // TODO: Test more extensively.
 test('extractStrings', async t => {
-  const response = await t.context.exampleQuery('extractStrings_0', build(e =>
-    e.extractStrings(e.data(d =>
+  const response = await t.context.exampleQuery('extractStrings_0',
+    e.extractStrings(e.data(
       d.struct({
         a: d.int8(1),
         b: d.string('foo'),
         c: d.struct({
-          list: d.list(
+          list: d.list([
             d.string('bar'),
-            d.string('baz')
-          ),
+            d.string('baz'),
+          ]),
         })
-      })
+      }).toDataConstructor()
     ))
-  ))
+  )
 
   t.deepEqual(response, ['bar', 'baz', 'foo'])
 })
 
 test('field', async t => {
-  const response = await t.context.exampleQuery('field_0', build(e =>
-    e.field('a', e.data(d => d.struct({
-        a: d.int8(1),
-        b: d.int8(2),
-        c: d.int8(3)}
-      ))
+  const response = await t.context.exampleQuery('field_0',
+    e.field('a',
+      e.data(d.struct({
+          a: d.int8(1),
+          b: d.int8(2),
+          c: d.int8(3)
+        }).toDataConstructor()
+      )
     )
-  ))
+  )
 
   t.is(response, 1)
 })
 
 test('setField', async t => {
-  const response = await t.context.exampleQuery('setField_0', build(e =>
-    e.setField('d', e.int8(4), e.data(d => d.struct({
-        a: d.int8(1),
-        b: d.int8(2),
-        c: d.int8(3)}
-      ))
+  const response = await t.context.exampleQuery('setField_0',
+    e.setField('d', e.int8(4), e.data(
+      d.struct({
+          a: d.int8(1),
+          b: d.int8(2),
+          c: d.int8(3)
+        }
+      ).toDataConstructor())
     )
-  ))
+  )
 
   t.deepEqual(response, {
     a: 1,
