+++ src/queries/expressions/union.ts
@@ -1,13 +1,24 @@
-import { buildExpression as build } from '@karma.run/sdk'
 import test from '../_before'
 
-test('isCase', async t => {
-  const response = await t.context.exampleQuery('isCase_0', build(e =>
-    e.data(d => d.list(
-      d.expr(e => e.isCase(e.data(d => d.union('foo', d.string('bar'))), e.string('foo'))),
-      d.expr(e => e.isCase(e.data(d => d.union('foo', d.string('bar'))), e.string('bar')))
-    ))
-  ))
+import * as e from '@karma.run/sdk/expression'
+import * as d from '@karma.run/sdk/value'
 
-  t.deepEqual(response, [true, false])
+test('isCase', async t => {
+  let response = await t.context.exampleQuery('isCase_0',
+    e.isCase(
+      e.string('foo'),
+      e.data(d.union('foo', d.string('bar')).toDataConstructor()),
+    )
+  )
+
+  t.true(response)
+
+  response = await t.context.exampleQuery('isCase_1',
+    e.isCase(
+      e.string('bar'),
+      e.data(d.union('foo', d.string('bar')).toDataConstructor()),
+    )
+  )
+
+  t.false(response)
 })
